Rename keyring to rnp_key_store first step -- the files
