+++ src/lib/create.c
@@ -81,7 +81,7 @@ __RCSID("$NetBSD: create.c,v 1.38 2010/11/15 08:03:39 agc Exp $");
 
 #include "bn.h"
 #include "create.h"
-#include "keyring_pgp.h"
+#include "key_store_pgp.h"
 #include "packet.h"
 #include "signature.h"
 #include "s2k.h"
