+++ src/lib/key_store.h
