+++ src/lib/key_store_pgp.c
@@ -79,7 +79,7 @@ __RCSID("$NetBSD: keyring.c,v 1.50 2011/06/25 00:37:44 agc Exp $");
 #endif
 
 #include "types.h"
-#include "keyring_pgp.h"
+#include "key_store_pgp.h"
 #include "packet-parse.h"
 #include "signature.h"
 #include "rnpsdk.h"
@@ -91,8 +91,8 @@ __RCSID("$NetBSD: keyring.c,v 1.50 2011/06/25 00:37:44 agc Exp $");
 #include "rnpdefs.h"
 #include "rnpdigest.h"
 #include <json.h>
-#include "keyring.h"
-#include "keyring_internal.h"
+#include "key_store.h"
+#include "key_store_internal.h"
 #include "packet-key.h"
 
 #include <sys/types.h>
