+++ src/lib/packet-parse.c
@@ -83,7 +83,7 @@ __RCSID("$NetBSD: packet-parse.c,v 1.51 2012/03/05 02:20:18 christos Exp $");
 #include "packet-parse.h"
 #include "packet-print.h"
 #include "packet-key.h"
-#include "keyring_pgp.h"
+#include "key_store_pgp.h"
 #include "errors.h"
 #include "packet-show.h"
 #include "create.h"
