+++ src/lib/writer.h
@@ -59,7 +59,7 @@
 #include "packet.h"
 #include "crypto.h"
 #include "errors.h"
-#include "keyring_pgp.h"
+#include "key_store_pgp.h"
 
 /**
  * \ingroup Writer
